lambda-function/requirements.txt
boto3
pillow
python-dotenv
line-bot-sdk
